enhance: useDebounce() returns [val, isPending]
