+++ packages/react/src/hooks/__tests__/useDebounce.web.tsx
@@ -1,8 +1,7 @@
 import { Endpoint } from '@data-client/endpoint';
-import { NetworkError } from '@data-client/rest';
-import { renderHook, act } from '@data-client/test';
+import { act } from '@data-client/test';
 import { render, waitFor } from '@testing-library/react';
-import React, { ReactElement, StrictMode, useTransition } from 'react';
+import React from 'react';
 
 import AsyncBoundary from '../../components/AsyncBoundary';
 import DataProvider from '../../components/DataProvider';
@@ -15,48 +14,6 @@ describe('useDebounce()', () => {
   });
   afterAll(() => {
     jest.useRealTimers();
-  });
-
-  it('should not update until delay has passed', () => {
-    const { result, rerender } = renderHook(
-      ({ value }: { value: string }) => {
-        return useDebounce(value, 100);
-      },
-      { initialProps: { value: 'initial' } },
-    );
-    expect(result.current).toEqual(['initial', false]);
-    jest.advanceTimersByTime(10);
-    rerender({ value: 'next' });
-    rerender({ value: 'third' });
-    expect(result.current).toEqual(['initial', false]);
-    act(() => {
-      jest.advanceTimersByTime(100);
-    });
-    expect(result.current).toEqual(['third', false]);
-  });
-
-  it('should never update when updatable is false', () => {
-    const { result, rerender } = renderHook(
-      ({ value, updatable }: { value: string; updatable: boolean }) => {
-        return useDebounce(value, 100, updatable);
-      },
-      { initialProps: { value: 'initial', updatable: false } },
-    );
-    expect(result.current).toEqual(['initial', false]);
-    jest.advanceTimersByTime(10);
-    rerender({ value: 'next', updatable: false });
-    act(() => {
-      jest.advanceTimersByTime(100);
-    });
-    expect(result.current).toEqual(['initial', false]);
-    rerender({ value: 'third', updatable: true });
-    expect(result.current).toEqual(['initial', false]);
-    jest.advanceTimersByTime(10);
-    expect(result.current).toEqual(['initial', false]);
-    act(() => {
-      jest.advanceTimersByTime(100);
-    });
-    expect(result.current).toEqual(['third', false]);
   });
 
   it('should be pending while async operation is performed based on debounced value', async () => {
